some fixes

+++ lib/iterable/iterable_double.dart
@@ -19,7 +19,7 @@ extension IterableOfDouble_ on Iterable<double> {
   /// [2.0, 4.0, 6.0, 8.0].average(); // 5.0
   /// ```
   double average() {
-    return this.averageBy((n) => n.toDouble());
+    return this.averageBy((n) => n);
   }
 
   /// Returns the largest value of all elements
